+++ scenario_runner.py
@@ -235,7 +235,7 @@ class ScenarioRunner(object):
                 self.ego_vehicles[i].set_transform(ego_vehicles[i].transform)
 
         # sync state
-        CarlaDataProvider.perform_carla_tick()
+        CarlaDataProvider.get_world().tick()
 
     def _analyze_scenario(self, config):
         """
@@ -285,9 +285,6 @@ class ScenarioRunner(object):
                             break
 
         self.world = self.client.get_world()
-        CarlaActorPool.set_client(self.client)
-        CarlaActorPool.set_world(self.world)
-        CarlaDataProvider.set_world(self.world)
 
         if self._args.agent:
             settings = self.world.get_settings()
@@ -295,9 +292,13 @@ class ScenarioRunner(object):
             settings.fixed_delta_seconds = 1.0 / self.frame_rate
             self.world.apply_settings(settings)
 
+        CarlaActorPool.set_client(self.client)
+        CarlaActorPool.set_world(self.world)
+        CarlaDataProvider.set_world(self.world)
+
         # Wait for the world to be ready
         if self.world.get_settings().synchronous_mode:
-            CarlaDataProvider.perform_carla_tick()
+            self.world.tick()
         else:
             self.world.wait_for_tick()