+++ Docs/CHANGELOG.md
@@ -23,6 +23,7 @@
     - Added new criteria, ActorSpeedAboveThresholdTest, useful to check if the ego vehicle has been standing still for long periods of time.
 * Setting up actors in batch now also randomizes their colors
 * When running routes, the weather parameters of each route can now be changed at will. Check the first route at srunner/data/routes_training.xml to see the correct format to do so. By default the weather is now a sunny midday.
+* Removed perform_carla_tick() function at CarlaDataProvider, which was a workaround for world.tick()
 * **Important** All challenge related content has been removed. This functionality has been improved and is now part of the [Leaderboard](https://github.com/carla-simulator/leaderboard). As a consequence:
     - The path to the autoagents has changed from .../challenge/autoagents to .../autoagents
     - The path to the route and scenario descriptions has changed from .../challenge to .../data